Add helper external program
Can run this external program to truncate large outputs in a vsh
buffer.  Can run it from either text editor.

As yet I have no "integration" between this program and the text
editors, one just has to know that it exists and to reference the path
directly (or, like I do, symlink it into ~/bin).

Keeping it here mostly for myself TBH.  So I don't have to copy this
around separate to my vsh install.