Put quotes around EOF marker on cat heredoc conversion
While reading an interesting blog post (https://justine.lol/lex/) saw
that bash heredocs can quote the end marker and that avoids variable
interpolation when handling it.

I'm not about to write code to escape everything that one might `cat`
out to the screen, but such a simple adjustment of using single quotes
around the marker in order to avoid such a large amount of the problems
that this functionality has is a huge and easy win.