Avoid error in original vim on closing
On unloading a buffer we record in a global dictionary a mapping from
buffer name to job.  On deleting a buffer we use that global dictionary
to close the relevant job.

In original vim I'd left the old problem of treating a job as a number
(which works in neovim) in this codepath.  When changing this I see I
still need to have some default, and rather than abstract away this
function entirely choosing different defaults depending on whether using
vim or neovim I instead check the type of the value we removed as well
as the value.

This ends up working for original vim as well as neovim.