Channel buffer mapping uses ch_info(c)['id'] as hash
Until now I would use the channel itself as a hash, but this would be
hashed as a string of the entire state.  That would include a
description of whether the channel was open or closed, which meant
accessing the channel_buffer_mapping after closing a channel gave
different results to doing it before having closed it.

This was stopping the clearup of variables on the shutdown of a vsh job
process.