Canonicalise access of vsh_job
I'd originally written with neovim in mind.  With neovim treating a job
as a number was just fine.  When porting to vim did not end up with
things being particularly clean.  Here we tidy things up.

Doing that by introducing a tri-state for the "vsh_job" status.  v:null
is "there is no process here", v:false is "this job has finished".

This tri-state is returned by the new s:vsh_job_status function.  Most
of the time the options are v:null or v:true.  For vim there is a short
time window where one could return v:false (between the process being
terminated and s:subprocess_closed finishing its call to
s:clear_vsh_vars).  For neovim there should be no time window where one
could return v:false, but on the other hand a time window where we could
theoretically return v:true even though the process is still running.

This time windows are a bit theoretical, but good to remember just in
case my model of the world is wrong and the exit callback isn't
guaranteed to be the first thing running after a channel is closed.

Throughout the rest of the codebase I adjust checks against vsh_job to
check against this tri-state returning v:true.

We have to have a slight representation overload for "no job running".
This is represented by there being no `vsh_job` variable in the given
buffer *or* the value being v:null.  This is needed because
s:subprocess_closed can be called when in a different buffer and in that
case we can't remove the buffer-local variable.  Hence we set it to
v:null as a second way to represent the same thing.