Avoid a performance problem in vsh_outputlen
This function requires generality to be called from outside of the
buffer we're working on, but that generality is really hurting our
performance.

The general case is only needed in an edge case -- we're adding text
into the buffer, but the output marker has been deleted.
In other cases we know that the cursor is at or after the prompt that we
care about.  In those cases we can use `search()` which is drastically
faster.

Here we split the function into two branches (chosen based on an
argument) where we can use the fast path when possible.