Fix vsh BOLOverride
Was not requiring prompt to be at the start of the line.  Hence I
could match a prompt in the middle of the line and decide to move to the
location where that prompt would end if it were at the start.

I suspect there are quite a few places where I'm making this mistake.
On scanning I think I saw the same problems in `s:prompt_end`.

Would like to fix sometime later -- fixing would require checking, and
that checking is too time-consuming for a work day.