Various comments around possible improvements
<Cmd> pseudokey for mappings instead of passing mode through the
functions (new feature ... guess I'll wait until it's so ubiquitous that
it's silly to not use -- at that point the time required to apply would
not be so much of a barrier).

I suspect there are bugs in my regexp comparisons.  Added a comment to
remind myself.  Comment claims it would be a good idea to go through all
uses and check, but I'm much more likely to only fix when hitting a bug
and use this comment as a reminder about possible problems.

Discussions about making the editor command do something similar to
fugitive.  Really not sure whether this is a good idea.  It works really
well for fugitive because there's only one buffer, but I'm using this
script as a general $EDITOR.
- Could have two different scripts, one as git editor and one general?
- Could put in the big engineering effort to record which files are open
  "from the vsh session" and mark each as "done" in turn -- then once
  all are "done" we run our exit command.
  - This to match how emacs works with `emacsclient`.
    Honestly not sure that this is even the best approach.