setup_env.vsh mentions when TERM == dumb
Otherwise when something is all good it's not very clear.  Have had new
users of this plugin open this file first and think that nothing is
working because of a combination of the `if` condition not printing
anything and a bug in vim meaning that bash is non-interactive so the
prompt is not printing either.