Avoid subprocesses when not backed by file
This is mostly there to avoid attempting to start a subprocess "in the
directory" of a fugitive:// buffer.  When running something like
:Gvdiffsplit I don't have any need to start a new subprocess for the
`vsh-mode` buffer representing the current Git Index.

Hoping that it will also apply to things like `:Linediff` and netrw
`scp://` files, on the principle that it doesn't make sense to have
shell processes locally to represent these files.